move route handler into serve function
